refactor(canard): set fractal maxDepth to 14 and dissolve micro-fractals beyond it (#331)
Set structure.maxDepth: 14 in .filid/config.json — the filid default (10)
predates this monorepo's 4-segment package prefix; 14 covers the blessed
fractal -> utils organ -> tests shape at the deepest legitimate fractals
(measured max 13) with one level of headroom. Violation fires above the
limit (depth > maxDepth).

Demote the two nodes that exceeded even that: extractConditionInfo and
getSimpleEquality shrink from single-function micro-fractals (depth 14,
own utils/__tests__ down to depth 16) into plain organ files under
getConditionIndexFactory/utils (depth 13). Their sole consumer is
getConditionIndexFactory, so per filid_code-placement §1 they default to
an organ of that owner; load-bearing invariants moved into the owner's
INTENT, the test joins ComputedPropertiesManager/utils/__tests__, and
the derivable dependency listing in that INTENT is compressed.

Verified: repo max directory depth 13, none above 14; resolveMaxDepth
returns 14 from the new config; typecheck/lint exit 0; package suite
3568/3568; doc detectors clean; structure_validate findings unchanged.
