--- packages/canard/schema-form/src/core/nodes/AbstractNode/utils/getComputedPropertiesManager/ComputedPropertiesManager/utils/getConditionIndexFactory/utils/extractConditionInfo/index.ts
@@ -1,1 +1,0 @@
-export { extractConditionInfo } from './extractConditionInfo';
