--- packages/canard/schema-form/src/core/nodes/AbstractNode/utils/getComputedPropertiesManager/ComputedPropertiesManager/utils/getConditionIndexFactory/utils/getSimpleEquality/index.ts
@@ -1,1 +1,0 @@
-export { getSimpleEquality } from './getSimpleEquality';
